+++ src/js/discharge-wizard/routes.jsx
@@ -1,9 +1,14 @@
-import DischargeWizardApp from './containers/DischargeWizardApp';
+import DischargeWizardApp from './DischargeWizardApp';
 import GuidancePage from './containers/GuidancePage';
+import FormPage from './containers/FormPage';
 
-const routes = [
-  { path: '/', component: DischargeWizardApp },
-  { path: '/guidance', component: GuidancePage }
-];
+const routes = {
+  path: '/',
+  component: DischargeWizardApp,
+  indexRoute: { component: FormPage },
+  childRoutes: [
+    { path: 'guidance', component: GuidancePage },
+  ],
+};
 
 export default routes;
